feaure: chage to use toRoute and fix module
