+++ core/domain/build.gradle.kts
@@ -2,7 +2,6 @@ plugins {
     id("hiraganaconverter.android.library")
     id("hiraganaconverter.android.library.jacoco")
     id("hiraganaconverter.android.hilt")
-    alias(libs.plugins.kotlin.serialization)
 }
 
 android {
