+++ core/ui/src/main/java/ksnd/hiraganaconverter/core/ui/navigation/NavKey.kt
@@ -1,4 +1,4 @@
-package ksnd.hiraganaconverter.core.domain
+package ksnd.hiraganaconverter.core.ui.navigation
 
 object NavKey {
     const val RECEIVED_TEXT = "received_text"
