+++ feature/converter/src/test/java/ksnd/hiraganaconverter/feature/converter/ConvertViewModelTest.kt
@@ -7,11 +7,11 @@ import io.mockk.coVerify
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import ksnd.hiraganaconverter.core.analytics.MockAnalyticsHelper
-import ksnd.hiraganaconverter.core.domain.NavKey
 import ksnd.hiraganaconverter.core.domain.usecase.ConversionFailedException
 import ksnd.hiraganaconverter.core.domain.usecase.ConvertTextUseCase
 import ksnd.hiraganaconverter.core.model.ui.HiraKanaType
 import ksnd.hiraganaconverter.core.testing.MainDispatcherRule
+import ksnd.hiraganaconverter.core.ui.navigation.NavKey
 import org.junit.Rule
 import org.junit.Test
